software/requirements.txt
pymodbus
flask
pyyaml
